Update endpoints.md
Minor grammatical changes for concision. Only 1 typo in the whole thing! One word I wasn't sure about is where you have byte\[] in the documentation, so your users see "[...] even specify the byte[] data for the content to be returned." I wasn't sure if that's the intended output. (Didn't change it.)