Update published-folders.md
This is in good shape. Tiny changes. One sentence is ungrammatical: "[...] it does so as the service account the process is running as." My idea is something like "it does so using the same service account as the process," but I'm not sure it's right anymore. I have a partial understanding here - I can view a customer's profile "as" them, but I'm not sure about the technical side of what's happening. We could talk it through tomorrow.