Update event-hubs.md
Almost no errors. Wasn't sure about a few things here. Like, "Event hubs are named and [...] " -- is this necessary? can we just tell them to create and name their event hub? or, wouldn't it be obvious that you have to give it a name when you create it?
Also, "This script is read into memory and not from disk" confused me because I was expecting the second part of the sentence to say "and not into __". So I didn't get why it wasn't "from" in both parts or "into" in both parts. If this isn't a real comparison, maybe we make them separate sentences or something. You could help me with this one tomorrow.