Update schedules.md
grammatical changes for concision. only a couple actual typos.
question: in most of your documentation, you write about running something *as* a user but here, you use the preposition *under*. Would it be correct to say *as* in the following sentence? "You can also define under which user the scheduled job runs".